Create lab 2 Q # 04.c
ctype.h is a standard C library header that provides functions and macros for character handling and classification. The header name "ctype" stands for "character type." This header is commonly used in C programs to perform various operations on characters, such as checking if a character is a digit, determining if it's an uppercase or lowercase letter, converting characters between cases, and more.
tolower(int c): Converts the character c to lowercase.
toupper(int c): Converts the character c to uppercase.